requirements-dev.txt
asgiref==3.4.1
Django==3.2.9
django-bootstrap-form==3.4
django-ckeditor==6.1.0
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-resized==0.3.11
libgravatar==1.0.0
Pillow==8.4.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
